requirements.txt
streamlit
pandas
pyyaml
numpy
plotly
pytest
plaid-python